requirements.txt
espn-api==0.36.0
requests==2.32.3
pytz==2024.1
python-dotenv==1.0.1
